refactor: remove comments, extract setup test storage
