+++ Cargo.toml
@@ -27,4 +27,5 @@ identity_stronghold = { version = "1.2.0", features = ["send-sync-storage"] }
 iota-sdk = { version = "1.0", features = ["stronghold"] }
 log = "0.4"
 signature = "2.2"
+test-log = "0.2"
 tokio = { version = "1", features = ["full"] }
