+++ did_jwk/src/consumer.rs
@@ -5,7 +5,7 @@ use log::info;
 use ssi_dids::did_resolve::ResolutionInputMetadata;
 use ssi_dids::DIDMethod;
 
-pub async fn resolve_did_jwk(did: CoreDID) -> std::result::Result<CoreDocument, identity_iota::core::Error> {
+pub async fn resolve_did_jwk(did: CoreDID) -> Result<CoreDocument, identity_iota::core::Error> {
     info!("Resolving DID: {}", did);
     let resolver = did_jwk_extern::DIDJWK.to_resolver();
     let input_metadata = ResolutionInputMetadata::default();
@@ -14,13 +14,9 @@ pub async fn resolve_did_jwk(did: CoreDID) -> std::result::Result<CoreDocument, 
     if let Some(error) = result.error.clone() {
         info!("Error: {:?}", error);
         return Err(identity_iota::core::Error::OneOrSetEmpty);
-        // return Err(Error::other(error));
     }
 
     info!("result: {:#?}", result);
-    // let key = resolve(did.as_str()).unwrap();
-    // info!("key: {:?}", key.fingerprint());
-    // let document = key.get_did_document(Config::default());
     info!("document: {:#?}", document);
     info!("metadata: {:#?}", metadata);
     CoreDocument::from_json(&document.to_json().unwrap())
