+++ did_web/Cargo.toml
@@ -25,4 +25,5 @@ urlencoding = "2.1"
 iota_stronghold = { version = "2.0" }
 rand = "0.8.5"
 shared = { path = "../shared", features = ["test"] }
+test-log.workspace = true
 wiremock = "0.5"
