+++ did_web/src/producer.rs
@@ -16,11 +16,7 @@ pub async fn produce_did_web(
     host: url::Host,
     port: Option<u16>,
 ) -> std::result::Result<CoreDocument, Error> {
-    // let exists = storage.key_storage().exists(key_id).await.unwrap();
-
-    // if !exists {
-    //     return Err(Error::other(format!("Key with id=[{}] does not exist", key_id)));
-    // }
+    // TODO: check if key exists for given key_id?
 
     let public_key_jwk = match storage {
         JwkStorageWrapper::Stronghold(stronghold_storage) => stronghold_storage.get_public_key(key_id).await.unwrap(),
@@ -35,7 +31,7 @@ pub async fn produce_did_web(
     url.set_host(Some(&host.to_string())).unwrap();
     url.set_port(port).unwrap();
 
-    info!("URL: {}", url.as_str());
+    info!("HOST: {}", url.as_str());
 
     let host_port = if port.is_some() {
         format!("{}:{}", url.host_str().unwrap(), url.port().unwrap())
@@ -43,41 +39,23 @@ pub async fn produce_did_web(
         url.host_str().unwrap().to_string()
     };
 
-    // let host_port = format!("{}:{}", url.host_str().unwrap(), url.port().unwrap_or_default());
     let host_port_encoded = urlencoding::encode(&host_port);
 
     let did_str = format!("did:web:{}", host_port_encoded);
 
-    // let did_str = match port {
-    //     Some(p) => format!("did:web:{}%3A{}", host, p),
-    //     None => format!("did:web:{}", host),
-    // };
-
-    // let url = url::Url::parse(&format!("https://{}/.well-known/did.json", host)).unwrap();
-    // info!("{}", url.port().unwrap());
-
-    // if let Some(did_str) = did_web_extern::DIDWeb.generate(&Source::Key(&jwk)) {
     info!("DID: {:?}", did_str);
 
     let controller = CoreDID::parse(&did_str).unwrap();
 
-    // info!("Controller: {:?}", controller.method_id());
-
     let verification_method =
         VerificationMethod::new_from_jwk(controller.clone(), public_key_jwk.clone(), Some("key-0")).unwrap();
-
-    // let assertion_method =
-    //     VerificationMethod::builder(BTreeMap::from([("foo".to_string(), "bar".to_json_value().unwrap())]))
-    //         .id(DIDUrl::new(did, url))
-    //         .build()
-    //         .unwrap();
 
     let mut properties = Object::new();
     properties.insert(
         "@context".to_string(),
         json!([
             "https://www.w3.org/ns/did/v1",
-            "https://w3id.org/security/suites/ed25519-2020/v1"
+            "https://w3id.org/security/suites/ed25519-2020/v1" // TODO: make dynamic
         ]),
     );
 
@@ -88,12 +66,12 @@ pub async fn produce_did_web(
         .build()
         .unwrap();
 
-    // TODO: Add "@context" to the document
-
-    info!("Host the following json under the following url:\n================================================");
+    info!("Host the following json under the following url:");
+    info!("================================================");
     info!("{}", url.join(".well-known/did.json").unwrap());
     info!("================================================");
     info!("{}", document.to_json_pretty().unwrap());
+    info!("================================================");
 
     Ok(document)
 }
@@ -106,35 +84,14 @@ mod tests {
 
     use identity_iota::core::ToJson;
     use identity_iota::did::DID;
-    use identity_iota::storage::JwkStorage;
-    use identity_stronghold::StrongholdStorage;
-    use iota_sdk::client::secret::stronghold::StrongholdSecretManager;
-    use iota_sdk::client::Password;
-    use shared::test_utils::{random_stronghold_path, test_jwk};
+    use shared::test_utils::new_stronghold_storage;
+    use test_log::test;
     use wiremock::matchers::{method, path};
     use wiremock::{Mock, MockServer, ResponseTemplate};
 
-    /// Issue: https://github.com/iotaledger/identity.rs/issues/1299
-    #[tokio::test]
+    #[test(tokio::test)]
     async fn produces_did_web() {
-        iota_stronghold::engine::snapshot::try_set_encrypt_work_factor(0).unwrap();
-
-        info!("====== Preparing Stronghold");
-
-        // Create stronghold
-        let stronghold = StrongholdSecretManager::builder()
-            .password(Password::from("secure_password".to_owned()))
-            .build(random_stronghold_path())
-            .unwrap();
-
-        let stronghold_storage = StrongholdStorage::new(stronghold);
-
-        // Generate key
-        let jwk = test_jwk();
-
-        // Insert into stronghold
-        let key_id = stronghold_storage.insert(jwk.clone()).await.unwrap();
-        info!("====== Done");
+        let (stronghold_storage, key_id) = new_stronghold_storage().await;
 
         // Start mock server and assert
         let mock_server = MockServer::start().await;
