+++ producer/Cargo.toml
@@ -24,7 +24,7 @@ identity_stronghold.workspace = true
 iota-sdk.workspace = true
 iota_stronghold = { version = "2.0" }
 log = "0.4"
-oid4vc-core = { git = "https://git@github.com/impierce/openid4vc.git", rev = "10a6bd7" }
+oid4vc-core = { git = "https://git@github.com/impierce/openid4vc.git", rev = "3fb6045" }
 rand = "0.8"
 serde = "1.0"
 serde_json = "1.0"
