+++ shared/Cargo.toml
@@ -13,5 +13,7 @@ test = []
 [dependencies]
 identity_iota.workspace = true
 identity_stronghold.workspace = true
+iota-sdk.workspace = true
+iota_stronghold = { version = "2.0" }
 log.workspace = true
 rand = "0.8"
